# src/main/resources/openapi3_1.yaml
openapi: 3.1.0
info:
    title: "Shopping-Microservice"
    summary: "Microservicio de compras de productos"
    description: "En este microservicio se desarollaran todos los componentes que se crean necesario para proveer todos los recursos relacionados a las peticiones de de ventas de productos"
    contact:
        name: Bautista
        email: Batubasilio@hotmail.com
    version: 1.0.0
    license: 
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"

servers:
    - url: "http://localhost:{port}/api/{basePath}/shopping/"
      variables:
        port:
            enum:
                - '8084'
                - '8080'
            default: '8080'
        basePath:
            default: v1
tags:
    - name: "Details" 
    - name: "Sales"

components:
    schemas:
        DetailsDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                quantity:
                    type: integer
                    format: int32
                productId:
                    type: integer
                    fotmat: int31
                product:
                    type: object
                    $ref: '#/components/schemas/ProductDTO'
                amount:
                    type: number
                    format: float
                discount:
                    type: number
                    format: float
                sale:
                    type: object
                    $ref: '#/components/schemas/SaleDTO'
                

        DetailsRequestDTO:
            type: object
            properties:
                quantity:
                    type: integer
                    format: int32
                productId:
                    type: integer
                    fotmat: int31
                amount:
                    type: number
                    format: float
                discount:
                    type: number
                    format: float
                sale:
                    type: object
                    $ref: '#/components/schemas/SaleDTO'

        DetailsListDTO:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/DetailsDTO'
                    
        SaleDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                clientId:
                    type: integer
                    format: int32
                amount:
                    type: number
                    format: float
                dateCreated:
                    type: string
                    format: date
                Details:
                    type: array
                    items:
                        type: object
                        $ref: '#/components/schemas/DetailsDTO'
                client:
                    type: object
                    $ref: '#/components/schemas/ClientDTO'

        SaleRequestDTO:
            type: object
            properties:
                clientId:
                    type: integer
                    format: int32
                amount:
                    type: number
                    format: float
                dateCreated:
                    type: string
                    format: date


        SaleListDTO:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/SaleDTO'

        ClientDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                lastname:
                    type: string
                dni:
                    type: integer
                    format: int32
                    minLength: 12
                cuil:
                    type: integer
                    format: int32
                    minLength: 17
                email:
                    type: string
                    format: email
                telephone:
                    type: string
                birth:
                    type: string
                    format: date
                Address:
                    type: object
                    $ref: '#/components/schemas/AddressDTO'
        AddressDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                number:
                    type: integer
                    format: int32
                location:
                    type: object
                    $ref: '#/components/schemas/LocationDTO'

        LocationDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                postalCode:
                    type: integer
                    format: int32
                province:
                    type: object
                    $ref: '#/components/schemas/ProvinceDTO'

        ProvinceDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string

        ProductDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                    description: Nombre del producto
                price:
                    type: number
                    description: Precio del producto
                status:
                    type: string
                    default: ENABLE
                    enum:
                        - ENABLE
                        - DISABLE
                        - SUSPENDED
                brand:
                    type: object
                    description: Marca del producto
                    $ref: '#/components/schemas/BrandDTO'

                dateCreated:
                    type: string
                    format: date

        PriceDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                price:
                    type: number
                product:
                    type: object
                    $ref: '#/components/schemas/ProductDTO'
                dateCreated:
                    type: string
                    format: date

        BrandDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                    description: "Marca"

                dateCreated:
                    type: string
                    format: date

paths:
    /details:
        post:
            tags:
                - Details
            security:
                - bearerAuth: []
            summary: Crear detalle
            operationId: createDetail
            requestBody: 
              content: 
                application/json: 
                  schema:
                    $ref: '#/components/schemas/DetailsRequestDTO'
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailsDTO'
                    
        put:
            tags:
                - Details
            security:
                - bearerAuth: []
            summary: Modificar detalle
            operationId: updateDetail
            requestBody: 
              content: 
                application/json: 
                  schema:
                    $ref: '#/components/schemas/DetailsDTO'
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailsDTO'
        get:
            tags:
                - Details
            security:
                - bearerAuth: []
            summary: retrieve all details
            operationId: retrieveAllDetails
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailsListDTO'
    /details/{id}:  
        get:
            tags:
                - Details
            security:
                - bearerAuth: []
            summary: retrieve all details
            operationId: retrieveDetailById
            parameters:
                - name: id
                  in: path
                  description: id del producto
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailsDTO'

        delete:
            tags:
                - Details
            security:
                - bearerAuth: []
            summary: remove product by id
            operationId: removeDetailById
            parameters:
                - name: id
                  in: path
                  description: id del producto
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                message: "Detalle borrado"
    /details/sale/{id}:
        get:
            tags:
                - Details
            security:
                - bearerAuth: [ ]
            summary: retrieve all details
            operationId: retrieveDetailBySaleId
            parameters:
                - name: id
                  in: path
                  description: id de la venta
                  required: true
                  schema:
                      type: integer
                      format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DetailsListDTO'

    
    /sales:
        post:
            tags:
                - Sales
            security:
                - bearerAuth: []
            summary: Crear venta
            operationId: createSale
            requestBody: 
              content: 
                application/json: 
                  schema:
                    $ref: '#/components/schemas/SaleRequestDTO'
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SaleDTO'
                    
        put:
            tags:
                - Sales
            security:
                - bearerAuth: []
            summary: Modificar venta
            operationId: updateSale
            requestBody: 
              content: 
                application/json: 
                  schema:
                    $ref: '#/components/schemas/SaleDTO'
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SaleDTO'
        get:
            tags:
                - Sales
            security:
                - bearerAuth: []
            summary: retrieve all sales
            operationId: retrieveAllSales
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SaleListDTO'
    /sales/{id}:  
        get:
            tags:
                - Sales
            security:
                - bearerAuth: []
            summary: retrieve sale by id
            operationId: retrieveSaleById
            parameters:
                - name: id
                  in: path
                  description: id de la venta
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SaleDTO'

        delete:
            tags:
                - Sales
            security:
                - bearerAuth: []
            summary: remove product by id
            operationId: removeSaleById
            parameters:
                - name: id
                  in: path
                  description: id del producto
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                message:
                                    type: string
                                    format: text
                                    default: Venta eliminada
    /sales/client/{id}:
        get:
            tags:
                - Sales
            security:
                - bearerAuth: [ ]
            summary: retrieve sale by cliente id
            operationId: retrieveSaleByClientId
            parameters:
                - name: id
                  in: path
                  description: id del cliente
                  required: true
                  schema:
                      type: integer
                      format: int32
            responses:
                200:
                    description: Si la consulta no falla
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SaleListDTO'


        
        
   